feat(MyBatis): 调整部分模型属性

+++ koala-infrastructures/koala-mybatis/src/main/java/cn/koala/mybatis/SortModel.java
@@ -11,12 +11,12 @@ public interface SortModel {
    *
    * @return 排序索引
    */
-  Integer getSortIndex();
+  Long getSortIndex();
 
   /**
    * 设置排序索引
    *
    * @param sortIndex 排序索引
    */
-  void setSortIndex(Integer sortIndex);
+  void setSortIndex(Long sortIndex);
 }
